Done with testing in the intro section.
Skipping validation in the notes, as it did not contain any code, only
theory, and as its code will be discussed in a section of its own.

There are a few issues right now. One, the reactive form is not working,
as before. And two, the tests are failing. Both these issues should be
resolved when doing the next section dedicated on Reactive Forms.